+++ resources/libraries/get_orgs.rb
@@ -1,19 +1,16 @@
-module Rb_manager
+module RbManager
   module Helpers
-    def get_orgs()
-
+    def get_orgs
       organizations = []
 
-      Chef::Role.list.keys.each do |m_key|
+      Chef::Role.list.each_key do |m_key|
         m = Chef::Role.load m_key
-        if !m.override_attributes["redborder"].nil? and !m.override_attributes["redborder"]["organization_uuid"].nil? and m.override_attributes["redborder"]["sensor_uuid"] == m.override_attributes["redborder"]["organization_uuid"]
-          organizations << m
-        end
+        next unless m.override_attributes['redborder'] && m.override_attributes['redborder']['organization_uuid'] && m.override_attributes['redborder']['sensor_uuid'] == m.override_attributes['redborder']['organization_uuid']
+
+        organizations << m
       end
 
       organizations
-
     end
   end
 end
-