+++ resources/libraries/managers_per_service.rb
@@ -1,19 +1,20 @@
 class Chef
   class Recipe
-    def managers_per_service()
-      services = node["redborder"]["services"]
-      cluster_info = node["redborder"]["cluster_info"] 
+    def managers_per_service
+      services = node['redborder']['services']
+      cluster_info = node['redborder']['cluster_info']
       cluster_services = {}
 
-      services.each do |serv, status|
+      services.each do |serv, _status|
         cluster_services[serv] = []
         cluster_info.each do |manager, info|
-          if !info["services"].nil? and info["services"].include?(serv)
-            cluster_services[serv] << manager 
+          if info['services'] && info['services'].include?(serv)
+            cluster_services[serv] << manager
           end
         end
       end
-      return cluster_services
+
+      cluster_services
     end
   end
 end