+++ resources/recipes/configure.rb
@@ -1,278 +1,367 @@
-#
-# Cookbook Name:: manager
+# Cookbook:: manager
 # Recipe:: configure
-#
-# Copyright 2024, redborder
-#
-# AFFERO GENERAL PUBLIC LICENSE V3
-#
+# Copyright:: 2024, redborder
+# License:: Affero General Public License, Version 3
 
 # Services configuration
 
 # manager services
 manager_services = manager_services()
 
-rb_selinux_config "Configure Selinux" do
-  if shell_out("getenforce").stdout.chomp == "Disabled"
+rb_common_config 'Configure common' do
+  action :configure
+end
+
+rb_selinux_config 'Configure Selinux' do
+  if shell_out('getenforce').stdout.chomp == 'Disabled'
     action :remove
   else
     action :add
   end
 end
 
-consul_config "Configure Consul Server" do
-  confdir node["consul"]["confdir"]
-  datadir node["consul"]["datadir"]
-  ipaddress node["ipaddress_sync"]
-  cdomain node["redborder"]["cdomain"]
-  dns_local_ip node["consul"]["dns_local_ip"]
-  (manager_services["consul"] ? (is_server true) : (is_server false))
-  action ((manager_services["consul"] or manager_services["consul-client"]) ? :add : :remove)
+consul_config 'Configure Consul Server' do
+  confdir node['consul']['confdir']
+  datadir node['consul']['datadir']
+  ipaddress node['ipaddress_sync']
+  cdomain node['redborder']['cdomain']
+  dns_local_ip node['consul']['dns_local_ip']
+  (manager_services['consul'] ? (is_server true) : (is_server false))
+  if manager_services['consul'] || manager_services['consul-client']
+    action :add
+  else
+    action :remove
+  end
 end
 
-if manager_services["chef-server"]
-  chef_server_config "Configure chef services" do
-    memory node["redborder"]["memory_services"]["chef-server"]["memory"]
-    postgresql false
-    postgresql_memory node["redborder"]["memory_services"]["postgresql"]["memory"]
-    chef_active manager_services["chef-server"]
-    ipaddress node["ipaddress_sync"]
+chef_server_config 'Configure chef services' do
+  memory node['redborder']['memory_services']['chef-server']['memory']
+  postgresql false
+  postgresql_memory node['redborder']['memory_services']['postgresql']['memory']
+  chef_active manager_services['chef-server']
+  ipaddress node['ipaddress_sync']
+  if manager_services['chef-server']
     action [:add, :register]
-  end
-else
-  chef_server_config "Remove chef service" do
+  else
     action [:remove, :deregister]
   end
 end
 
-zookeeper_config "Configure Zookeeper" do
-  port node["zookeeper"]["port"]
-  memory node["redborder"]["memory_services"]["zookeeper"]["memory"]
-  hosts node["redborder"]["managers_per_services"]["zookeeper"]
-  ipaddress node["ipaddress_sync"]
-  action (manager_services["zookeeper"] ? [:add, :register] : [:remove, :deregister])
+zookeeper_config 'Configure Zookeeper' do
+  port node['zookeeper']['port']
+  memory node['redborder']['memory_services']['zookeeper']['memory']
+  hosts node['redborder']['managers_per_services']['zookeeper']
+  ipaddress node['ipaddress_sync']
+  if manager_services['zookeeper']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-kafka_config "Configure Kafka" do
-  memory node["redborder"]["memory_services"]["kafka"]["memory"]
-  maxsize node["redborder"]["manager"]["hd_services_current"]["kafka"]
-  managers_list node["redborder"]["managers_per_services"]["kafka"]
-  zk_hosts node["redborder"]["zookeeper"]["zk_hosts"]
-  host_index node["redborder"]["kafka"]["host_index"]
-  ipaddress node["ipaddress_sync"]
-  action (manager_services["kafka"] ? [:add, :register] : [:remove, :deregister])
+kafka_config 'Configure Kafka' do
+  memory node['redborder']['memory_services']['kafka']['memory']
+  maxsize node['redborder']['manager']['hd_services_current']['kafka']
+  managers_list node['redborder']['managers_per_services']['kafka']
+  zk_hosts node['redborder']['zookeeper']['zk_hosts']
+  host_index node['redborder']['kafka']['host_index']
+  ipaddress node['ipaddress_sync']
+  if manager_services['kafka']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-if  manager_services["druid-coordinator"] or
-    manager_services["druid-overlord"] or
-    manager_services["druid-broker"] or
-    manager_services["druid-middlemanager"] or
-    manager_services["druid-historical"] or
-    manager_services["druid-realtime"]
-
-  ["druid-broker", "druid-coordinator", "druid-historical",
-  "druid-middlemanager", "druid-overlord"].each do |druid_service|
+if manager_services['druid-coordinator'] || manager_services['druid-overlord'] || manager_services['druid-broker'] || manager_services['druid-middlemanager'] || manager_services['druid-historical'] || manager_services['druid-realtime']
+  %w(druid-broker druid-coordinator druid-historical
+  druid-middlemanager druid-overlord).each do |druid_service|
     service druid_service do
-      supports :status => true, :start => true, :restart => true, :reload => true
+      supports status: true, start: true, restart: true, reload: true
       action :nothing
     end
   end
 
-  druid_common "Configure druid common resources" do
-    name node["hostname"]
-    zookeeper_hosts node["redborder"]["zookeeper"]["zk_hosts"]
-    memcached_hosts node["redborder"]["memcached"]["hosts"]
-    s3_service "s3.service"
-    s3_port node["minio"]["port"]
-    cdomain node["redborder"]["cdomain"]
+  druid_common 'Configure druid common resources' do
+    name node['hostname']
+    zookeeper_hosts node['redborder']['zookeeper']['zk_hosts']
+    memcached_hosts node['redborder']['memcached']['hosts']
+    s3_service 's3.service'
+    s3_port node['minio']['port']
+    cdomain node['redborder']['cdomain']
     action :add
-    notifies :restart, 'service[druid-broker]', :delayed if manager_services["druid-broker"]
-    notifies :restart, 'service[druid-coordinator]', :delayed if manager_services["druid-coordinator]"]
-    notifies :restart, 'service[druid-historical]', :delayed if manager_services["druid-historical"]
-    notifies :restart, 'service[druid-middlemanager]', :delayed if manager_services["druid-middlemanager"]
-    notifies :restart, 'service[druid-overlord]', :delayed if manager_services["druid-overlord"]
+    notifies :restart, 'service[druid-broker]', :delayed if manager_services['druid-broker']
+    notifies :restart, 'service[druid-coordinator]', :delayed if manager_services['druid-coordinator]']
+    notifies :restart, 'service[druid-historical]', :delayed if manager_services['druid-historical']
+    notifies :restart, 'service[druid-middlemanager]', :delayed if manager_services['druid-middlemanager']
+    notifies :restart, 'service[druid-overlord]', :delayed if manager_services['druid-overlord']
   end
 else
-  druid_common "Delete druid common resources" do
+  druid_common 'Delete druid common resources' do
     action :remove
   end
 end
 
-druid_coordinator "Configure Druid Coordinator" do
-  name node["hostname"]
-  ipaddress node["ipaddress_sync"]
-  memory_kb node["redborder"]["memory_services"]["druid-coordinator"]["memory"]
-  action (manager_services["druid-coordinator"] ? [:add, :register] : [:remove, :deregister])
-end
-
-druid_overlord "Configure Druid Overlord" do
-  name node["hostname"]
-  ipaddress node["ipaddress_sync"]
-  memory_kb node["redborder"]["memory_services"]["druid-overlord"]["memory"]
-  action (manager_services["druid-overlord"] ? [:add, :register] : [:remove, :deregister])
-end
-
-druid_broker "Configure Druid Broker" do
-  name node["hostname"]
-  ipaddress node["ipaddress_sync"]
-  memory_kb node["redborder"]["memory_services"]["druid-broker"]["memory"]
-  action (manager_services["druid-broker"] ? [:add, :register] : [:remove, :deregister])
-end
-
-druid_middlemanager "Configure Druid MiddleManager" do
-  name node["hostname"]
-  ipaddress node["ipaddress_sync"]
-  memory_kb node["redborder"]["memory_services"]["druid-middlemanager"]["memory"]
-  action (manager_services["druid-middlemanager"] ? [:add, :register] : [:remove, :deregister])
-end
-
-druid_historical "Configure Druid Historical" do
-  name node["hostname"]
-  ipaddress node["ipaddress_sync"]
-  memory_kb node["redborder"]["memory_services"]["druid-historical"]["memory"]
-  action (manager_services["druid-historical"] ? [:add, :register] : [:remove, :deregister])
-end
-
-druid_realtime "Configure Druid Realtime" do
-  name node["hostname"]
-  ipaddress node["ipaddress_sync"]
-  zookeeper_hosts node["redborder"]["zookeeper"]["zk_hosts"]
-  partition_num node["redborder"]["druid"]["realtime"]["partition_num"]
-  memory_kb node["redborder"]["memory_services"]["druid-realtime"]["memory"]
-  action (manager_services["druid-realtime"] ? [:add, :register] : [:remove, :deregister])
-end
-
-memcached_config "Configure Memcached" do
-  memory node["redborder"]["memory_services"]["memcached"]["memory"]
-  ipaddress node["ipaddress_sync"]
-  action (manager_services["memcached"] ? [:add, :register] : [:remove, :deregister])
-end
-
-mongodb_config "Configure Mongodb" do
-  action (manager_services["mongodb"] ? [:add, :register] : [:remove, :deregister])
-end
-
-if manager_services["hadoop-nodemanager"] or
-   manager_services["hadoop-resourcemanager"]
-
-  hadoop_common "Configure hadoop common resources" do
-    name node["hostname"]
-    zookeeper_hosts node["redborder"]["zookeeper"]["zk_hosts"]
-    memory_kb node["redborder"]["memory_services"]["hadoop-nodemanager"]["memory"]
-    containersMemory node["redborder"]["hadoop"]["containersMemory"]
-    action :add
+druid_coordinator 'Configure Druid Coordinator' do
+  name node['hostname']
+  ipaddress node['ipaddress_sync']
+  memory_kb node['redborder']['memory_services']['druid-coordinator']['memory']
+  if manager_services['druid-coordinator']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
   end
-else
-  hadoop_common "Delete hadoop common resources" do
+end
+
+druid_overlord 'Configure Druid Overlord' do
+  name node['hostname']
+  ipaddress node['ipaddress_sync']
+  memory_kb node['redborder']['memory_services']['druid-overlord']['memory']
+  if manager_services['druid-overlord']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
+end
+
+druid_broker 'Configure Druid Broker' do
+  name node['hostname']
+  ipaddress node['ipaddress_sync']
+  memory_kb node['redborder']['memory_services']['druid-broker']['memory']
+  if manager_services['druid-broker']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
+end
+
+druid_middlemanager 'Configure Druid MiddleManager' do
+  name node['hostname']
+  ipaddress node['ipaddress_sync']
+  memory_kb node['redborder']['memory_services']['druid-middlemanager']['memory']
+  if manager_services['druid-middlemanager']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
+end
+
+druid_historical 'Configure Druid Historical' do
+  name node['hostname']
+  ipaddress node['ipaddress_sync']
+  memory_kb node['redborder']['memory_services']['druid-historical']['memory']
+  if manager_services['druid-historical']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
+end
+
+druid_realtime 'Configure Druid Realtime' do
+  name node['hostname']
+  ipaddress node['ipaddress_sync']
+  zookeeper_hosts node['redborder']['zookeeper']['zk_hosts']
+  partition_num node['redborder']['druid']['realtime']['partition_num']
+  memory_kb node['redborder']['memory_services']['druid-realtime']['memory']
+  if manager_services['druid-realtime']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
+end
+
+memcached_config 'Configure Memcached' do
+  memory node['redborder']['memory_services']['memcached']['memory']
+  ipaddress node['ipaddress_sync']
+  if manager_services['memcached']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
+end
+
+mongodb_config 'Configure Mongodb' do
+  if manager_services['mongodb']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
+end
+
+hadoop_common 'Configure hadoop common resources' do
+  name node['hostname']
+  zookeeper_hosts node['redborder']['zookeeper']['zk_hosts']
+  memory_kb node['redborder']['memory_services']['hadoop-nodemanager']['memory']
+  containersMemory node['redborder']['hadoop']['containersMemory']
+  if manager_services['hadoop-nodemanager'] || manager_services['hadoop-resourcemanager']
+    action :add
+  else
     action :remove
   end
 end
 
-hadoop_resourcemanager "Configure Hadoop ResourceManager" do
-  memory_kb node["redborder"]["memory_services"]["hadoop-resourcemanager"]["memory"]
-  action (manager_services["hadoop-resourcemanager"] ? [:add, :register] : [:remove, :deregister])
+hadoop_resourcemanager 'Configure Hadoop ResourceManager' do
+  memory_kb node['redborder']['memory_services']['hadoop-resourcemanager']['memory']
+  if manager_services['hadoop-resourcemanager']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-hadoop_nodemanager "Configure Hadoop NodeManager" do
-  memory_kb node["redborder"]["memory_services"]["hadoop-nodemanager"]["memory"]
-  action (manager_services["hadoop-nodemanager"] ? [:add, :register] : [:remove, :deregister])
+hadoop_nodemanager 'Configure Hadoop NodeManager' do
+  memory_kb node['redborder']['memory_services']['hadoop-nodemanager']['memory']
+  if manager_services['hadoop-nodemanager']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-samza_config "Configure samza applications" do
-  memory_per_container node["redborder"]["samza"]["memory_per_container"]
-  num_containers node["redborder"]["samza"]["num_containers"]
-  action (manager_services["hadoop-nodemanager"] ? :add : :remove)
+samza_config 'Configure samza applications' do
+  memory_per_container node['redborder']['samza']['memory_per_container']
+  num_containers node['redborder']['samza']['num_containers']
+  if manager_services['hadoop-nodemanager']
+    action :add
+  else
+    action :remove
+  end
 end
 
-geoip_config "Configure GeoIP" do
-  action (manager_services["geoip"] ? :add : :remove)
+geoip_config 'Configure GeoIP' do
+  action :add
 end
 
-snmp_config "Configure snmp" do
-  hostname node["hostname"]
-  cdomain node["redborder"]["cdomain"]
-  action (manager_services["snmp"] ? :add : :remove)
+snmp_config 'Configure snmp' do
+  hostname node['hostname']
+  cdomain node['redborder']['cdomain']
+  if manager_services['snmp']
+    action :add
+  else
+    action :remove
+  end
 end
 
-rbmonitor_config "Configure redborder-monitor" do
-  name node["hostname"]
-  device_nodes node["redborder"]["sensors_info_all"]["device-sensor"]
-  flow_nodes node["redborder"]["sensors_info_all"]["flow-sensor"]
-  managers node["redborder"]["managers_list"]
-  cluster node["redborder"]["cluster_info"]
-  hostip node["redborder"]["cluster_info"][name]["ip"]
-  action (manager_services["redborder-monitor"] ? :add : :remove)
+rbmonitor_config 'Configure redborder-monitor' do
+  name node['hostname']
+  device_nodes node.run_state['sensors_info_all']['device-sensor']
+  flow_nodes node.run_state['sensors_info_all']['flow-sensor']
+  managers node['redborder']['managers_list']
+  cluster node['redborder']['cluster_info']
+  hostip node['redborder']['cluster_info'][name]['ip']
+  if manager_services['redborder-monitor']
+    action :add
+  else
+    action :remove
+  end
 end
 
-rbscanner_config "Configure redborder-scanner" do
-  scanner_nodes node["redborder"]["sensors_info_all"]["scanner-sensor"]
-  action (manager_services["redborder-scanner"] ? [:add, :register] : [:remove, :deregister])
+rbscanner_config 'Configure redborder-scanner' do
+  scanner_nodes node.run_state['sensors_info_all']['scanner-sensor']
+  if manager_services['redborder-scanner']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-nginx_config "Configure Nginx" do
-  cdomain node["redborder"]["cdomain"]
-  action (manager_services["nginx"] ? [:add, :register] : [:remove, :deregister])
+nginx_config 'Configure Nginx' do
+  cdomain node['redborder']['cdomain']
+  if manager_services['nginx']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-if manager_services["nginx"] and manager_services["chef-server"]
-  nginx_config "Configure Nginx Chef" do
-    service_name "erchef"
-    cdomain node["redborder"]["cdomain"]
+nginx_config 'Configure Nginx Chef' do
+  service_name 'erchef'
+  cdomain node['redborder']['cdomain']
+  if manager_services['nginx'] && manager_services['chef-server']
     action [:configure_certs, :add_erchef]
+  else
+    action :nothing
   end
 end
 
-if manager_services["nginx"] and manager_services["rb-aioutliers"]
-  nginx_config "Configure Nginx aioutliers" do
-    service_name "rb-aioutliers"
-    cdomain node["redborder"]["cdomain"]
+nginx_config 'Configure Nginx aioutliers' do
+  service_name 'rb-aioutliers'
+  cdomain node['redborder']['cdomain']
+  if manager_services['nginx'] && manager_services['rb-aioutliers']
     action [:configure_certs, :add_aioutliers]
+  else
+    action :nothing
   end
 end
 
-webui_config "Configure WebUI" do
-  hostname node["hostname"]
-  memory_kb node["redborder"]["memory_services"]["webui"]["memory"]
-  cdomain node["redborder"]["cdomain"]
-  port node["redborder"]["webui"]["port"]
-  action (manager_services["webui"] ? [:add, :register, :configure_rsa] : [:remove, :deregister])
+webui_config 'Configure WebUI' do
+  hostname node['hostname']
+  memory_kb node['redborder']['memory_services']['webui']['memory']
+  cdomain node['redborder']['cdomain']
+  port node['redborder']['webui']['port']
+  if manager_services['webui']
+    action [:add, :register, :configure_rsa]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-webui_config "Configure Nginx WebUI" do
-  cdomain node["redborder"]["cdomain"]
-  port node["redborder"]["webui"]["port"]
-  action ((manager_services["webui"] and manager_services["nginx"]) ? [:configure_certs, :add_webui_conf_nginx] : :nothing)
+webui_config 'Configure Nginx WebUI' do
+  hosts node['redborder']['webui']['hosts']
+  cdomain node['redborder']['cdomain']
+  port node['redborder']['webui']['port']
+  if manager_services['webui'] && manager_services['nginx']
+    action [:configure_certs, :add_webui_conf_nginx]
+  else
+    action :nothing
+  end
 end
 
-http2k_config "Configure Http2k" do
-  domain node["redborder"]["cdomain"]
-  kafka_hosts node["redborder"]["managers_per_services"]["kafka"]
-  memory node["redborder"]["memory_services"]["http2k"]["memory"]
-  port node["redborder"]["http2k"]["port"]
-  proxy_nodes node["redborder"]["sensors_info"]["proxy-sensor"]
-  ips_nodes node["redborder"]["sensors_info"]["ips-sensor"]
-  ipsg_nodes node["redborder"]["sensors_info"]["ipsg-sensor"]
-  ipscp_nodes node["redborder"]["sensors_info"]["ipscp-sensor"]
-  organizations node["redborder"]["organizations"]
-  locations_list node["redborder"]["locations"]
-  action (manager_services["http2k"]  ? [:add, :register] : [:remove, :deregister])
+http2k_config 'Configure Http2k' do
+  domain node['redborder']['cdomain']
+  kafka_hosts node['redborder']['managers_per_services']['kafka']
+  memory node['redborder']['memory_services']['http2k']['memory']
+  port node['redborder']['http2k']['port']
+  proxy_nodes node.run_state['sensors_info']['proxy-sensor']
+  ips_nodes node.run_state['sensors_info']['ips-sensor']
+  ipsg_nodes node.run_state['sensors_info']['ipsg-sensor']
+  ipscp_nodes node.run_state['sensors_info']['ipscp-sensor']
+  organizations node['redborder']['organizations']
+  locations_list node['redborder']['locations']
+  if manager_services['http2k']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-http2k_config "Configure Nginx Http2k" do
-  domain node["redborder"]["cdomain"]
-  port node["redborder"]["http2k"]["port"]
-  action ((manager_services["http2k"] and manager_services["nginx"]) ? [:configure_certs, :add_http2k_conf_nginx] : :nothing)
+http2k_config 'Configure Nginx Http2k' do
+  domain node['redborder']['cdomain']
+  port node['redborder']['http2k']['port']
+  if manager_services['http2k'] && manager_services['nginx']
+    action [:configure_certs, :add_http2k_conf_nginx]
+  else
+    action :nothing
+  end
 end
 
-f2k_config "Configure f2k" do
-  sensors node["redborder"]["sensors_info"]["flow-sensor"]
-  action (manager_services["f2k"] ? [:add, :register] : [:remove, :deregister])
+f2k_config 'Configure f2k' do
+  sensors node.run_state['sensors_info']['flow-sensor']
+  if manager_services['f2k']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-pmacct_config "Configure pmacct" do
-  sensors node["redborder"]["sensors_info"]["flow-sensor"]
-  kafka_hosts node["redborder"]["managers_per_services"]["kafka"]
-  action (manager_services["pmacct"] ? [:add, :register] : [:remove, :deregister])
+pmacct_config 'Configure pmacct' do
+  sensors node.run_state['sensors_info']['flow-sensor']
+  kafka_hosts node['redborder']['managers_per_services']['kafka']
+  if manager_services['pmacct']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
 logstash_config 'Configure logstash' do
@@ -283,168 +372,246 @@ logstash_config 'Configure logstash' do
   scanner_nodes node.run_state['sensors_info_all']['scanner-sensor']
   device_nodes node.run_state['sensors_info_all']['device-sensor']
   logstash_pipelines node.default['pipelines']
-  if manager_services['logstash'] && node.default['pipelines'] && !node.default['pipelines'].empty?
+  if manager_services['logstash'] && !node.default['pipelines'].nil? && !node.default['pipelines'].empty?
     action [:add, :register]
   else
     action [:remove, :deregister]
   end
 end
 
-rbdswatcher_config "Configure redborder-dswatcher" do
-  cdomain node["redborder"]["cdomain"]
-  action (manager_services["redborder-dswatcher"] ? [:add, :register] : [:remove, :deregister])
+rbdswatcher_config 'Configure redborder-dswatcher' do
+  cdomain node['redborder']['cdomain']
+  if manager_services['redborder-dswatcher']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-rbevents_counter_config "Configure redborder-events-counter" do
-  cdomain node["redborder"]["cdomain"]
-  action (manager_services["redborder-events-counter"] ? [:add, :register] : [:remove, :deregister])
+rbevents_counter_config 'Configure redborder-events-counter' do
+  cdomain node['redborder']['cdomain']
+  if manager_services['redborder-events-counter']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-rsyslog_config "Configure rsyslog" do
-  vault_nodes node["redborder"]["sensors_info_all"]["vault-sensor"] + node["redborder"]["sensors_info_all"]["cep-sensor"]
-  ips_nodes node["redborder"]["sensors_info_all"]["ips-sensor"] + node["redborder"]["sensors_info_all"]["ipsv2-sensor"] + node["redborder"]["sensors_info_all"]["ipscp-sensor"]
-  action (manager_services["rsyslog"] ? [:add, :register] : [:remove, :deregister])
+rsyslog_config 'Configure rsyslog' do
+  vault_nodes node.run_state['sensors_info_all']['vault-sensor'] + node.run_state['sensors_info_all']['cep-sensor']
+  ips_nodes node.run_state['sensors_info_all']['ips-sensor'] + node.run_state['sensors_info_all']['ipsv2-sensor'] + node.run_state['sensors_info_all']['ipscp-sensor']
+  if manager_services['rsyslog']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-rbnmsp_config "Configure redborder-nmsp" do
-  memory node["redborder"]["memory_services"]["redborder-nmsp"]["memory"]
-  proxy_nodes node["redborder"]["sensors_info_all"]["proxy-sensor"]
-  flow_nodes node["redborder"]["sensors_info_all"]["flow-sensor"]
-  hosts node["redborder"]["zookeeper"]["zk_hosts"]
-  action (manager_services["redborder-nmsp"] ? [:add, :configure_keys, :register] : [:remove, :deregister])
+rbnmsp_config 'Configure redborder-nmsp' do
+  memory node['redborder']['memory_services']['redborder-nmsp']['memory']
+  proxy_nodes node.run_state['sensors_info_all']['proxy-sensor']
+  flow_nodes node.run_state['sensors_info_all']['flow-sensor']
+  hosts node['redborder']['zookeeper']['zk_hosts']
+  if manager_services['redborder-nmsp']
+    action [:add, :configure_keys, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-n2klocd_config "Configure n2klocd" do
-  mse_nodes node["redborder"]["sensors_info_all"]["mse-sensor"]
-  meraki_nodes node["redborder"]["sensors_info_all"]["meraki-sensor"]
-  n2klocd_managers node["redborder"]["managers_per_services"]["n2klocd"]
-  memory node["redborder"]["memory_services"]["n2klocd"]["memory"]
-  action (manager_services["n2klocd"] ? [:add, :register] : [:remove, :deregister])
+n2klocd_config 'Configure n2klocd' do
+  mse_nodes node.run_state['sensors_info_all']['mse-sensor']
+  meraki_nodes node.run_state['sensors_info_all']['meraki-sensor']
+  n2klocd_managers node['redborder']['managers_per_services']['n2klocd']
+  memory node['redborder']['memory_services']['n2klocd']['memory']
+  if manager_services['n2klocd']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-rbale_config "Configure redborder-ale" do
-  ale_nodes node["redborder"]["sensors_info_all"]["ale-sensor"]
-  action (node["redborder"]["services"]["redborder-ale"] ? [:add, :register] : [:remove, :deregister])
+rbale_config 'Configure redborder-ale' do
+  ale_nodes node.run_state['sensors_info_all']['ale-sensor']
+  if node['redborder']['services']['redborder-ale']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-rblogstatter_config "Configure redborder-logstatter" do
-  action (node["redborder"]["services"]["rb-logstatter"] ? :add : :remove)
+rblogstatter_config 'Configure redborder-logstatter' do
+  if node['redborder']['services']['rb-logstatter']
+    action :add
+  else
+    action :remove
+  end
 end
 
-rb_arubacentral_config "Configure rb-arubacentral" do
-  arubacentral_nodes node["redborder"]["sensors_info_all"]["arubacentral-sensor"]
-  flow_nodes node["redborder"]["sensors_info_all"]["flow-sensor"]
-  action (node["redborder"]["services"]["rb-arubacentral"] ? :add : :remove)
+rb_arubacentral_config 'Configure rb-arubacentral' do
+  arubacentral_nodes node.run_state['sensors_info_all']['arubacentral-sensor']
+  flow_nodes node.run_state['sensors_info_all']['flow-sensor']
+  if node['redborder']['services']['rb-arubacentral']
+    action :add
+  else
+    action :remove
+  end
 end
 
-#freeradius_config "Configure radiusd" do
-#  flow_nodes node["redborder"]["sensors_info_all"]["flow-sensor"]
-#  action (node["redborder"]["services"]["radiusd"] ? [:config_common, :config_manager, :register] : [:remove, :deregister])
-#end
+# freeradius_config 'Configure radiusd' do
+#   flow_nodes node.run_state['sensors_info_all']['flow-sensor']
+#   action (node['redborder']['services']['radiusd'] ? [:config_common, :config_manager, :register] : [:remove, :deregister])
+# end
 
-rbaioutliers_config "Configure rb-aioutliers" do
-  action (manager_services["rb-aioutliers"] ? [:add, :register] : [:remove, :deregister])
+rbaioutliers_config 'Configure rb-aioutliers' do
+  if manager_services['rb-aioutliers']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-rbcep_config "Configure redborder-cep" do
-  flow_nodes node["redborder"]["sensors_info_all"]["flow-sensor"]
-  vault_nodes node["redborder"]["sensors_info_all"]["vault-sensor"]
-  ips_nodes node["redborder"]["sensors_info_all"]["ips-sensor"] + node["redborder"]["sensors_info_all"]["ipsv2-sensor"] + node["redborder"]["sensors_info_all"]["ipscp-sensor"]
-  action (node["redborder"]["services"]["redborder-cep"] ? [:add, :register] : [:remove, :deregister])
+rbcep_config 'Configure redborder-cep' do
+  flow_nodes node.run_state['sensors_info_all']['flow-sensor']
+  vault_nodes node.run_state['sensors_info_all']['vault-sensor']
+  ips_nodes node.run_state['sensors_info_all']['ips-sensor'] + node.run_state['sensors_info_all']['ipsv2-sensor'] + node.run_state['sensors_info_all']['ipscp-sensor']
+  if node['redborder']['services']['redborder-cep']
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-rbcgroup_config "Configure cgroups" do
+rb_postfix_config 'Configure postfix' do
+  if node['redborder']['services']['postfix']
+    action :add
+  else
+    action :remove
+  end
+end
+
+rbcgroup_config 'Configure cgroups' do
   action :add
 end
 
 # Determine external
-external_services = Chef::DataBagItem.load("rBglobal", "external_services")
-
-postgresql_config "Configure postgresql" do
-  cdomain node["redborder"]["cdomain"]
-  ipaddress node["ipaddress_sync"]
-  action (manager_services["postgresql"] and external_services["postgresql"] == "onpremise" ? [:add, :register] : [:remove, :deregister])
+begin
+  external_services = data_bag_item('rBglobal', 'external_services')
+rescue
+  external_services = {}
 end
 
-s3_leader = `serf members | grep s3=ready | awk '{print $1'} | head -n 1`.strip
-
-# Allow only one s3 onpremise node for now.. TODO: Distributed MinIO
-minio_config "Configure S3 (minio)" do
-  ipaddress node["ipaddress_sync"]
-  action ((manager_services["s3"] and external_services["s3"] == "onpremise" and s3_leader == node.name ) ? [:add, :register] : [:remove, :deregister])
+postgresql_config 'Configure postgresql' do
+  cdomain node['redborder']['cdomain']
+  ipaddress node['ipaddress_sync']
+  if manager_services['postgresql'] && external_services['postgresql'] == 'onpremise'
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
-if manager_services["s3"]
-  nginx_config "Configure S3 certs" do
-    service_name "s3"
-    cdomain node["redborder"]["cdomain"]
+begin
+  s3_secrets = data_bag_item('passwords', 's3')
+rescue
+  ssh_secrets = {}
+end
+
+# Allow only s3 onpremise nodes for now..
+minio_config 'Configure S3 (minio)' do
+  ipaddress node['ipaddress_sync']
+  access_key_id s3_secrets['s3_access_key_id']
+  secret_key_id s3_secrets['s3_secret_key_id']
+  action((manager_services['s3'] && (external_services['s3'] == 'onpremise')) ? [:add, :register] : [:remove, :deregister])
+end
+
+# First configure the cert for the service before configuring nginx
+if manager_services['s3']
+  nginx_config 'Configure S3 certs' do
+    service_name 's3'
+    cdomain node['redborder']['cdomain']
     action :configure_certs
   end
 end
 
-ssh_secrets = Chef::DataBagItem.load("passwords", "ssh") rescue ssh_secrets = {}
+# Configure Nginx s3 onpremise nodes for now..
+minio_config 'Configure Nginx S3 (minio)' do
+  s3_hosts node['redborder']['s3']['s3_hosts']
+  action((manager_services['s3'] && (external_services['s3'] == 'onpremise')) ? [:add_s3_conf_nginx] : :nothing)
+end
 
-directory "/root/.ssh" do
-  owner "root"
-  group "root"
-  mode 0755
+begin
+  ssh_secrets = data_bag_item('passwords', 'ssh')
+rescue
+  ssh_secrets = {}
+end
+
+directory '/root/.ssh' do
+  owner 'root'
+  group 'root'
+  mode '0755'
   action :create
 end
 
-if !ssh_secrets.empty?
-  template "/root/.ssh/authorized_keys" do
-    source "rsa.pub.erb"
-    owner "root"
-    group "root"
-    mode 0600
+unless ssh_secrets.empty?
+  template '/root/.ssh/authorized_keys' do
+    source 'rsa.pub.erb'
+    owner 'root'
+    group 'root'
+    mode '0600'
     retries 2
-    variables(:public_rsa => ssh_secrets['public_rsa'])
+    variables(public_rsa: ssh_secrets['public_rsa'])
   end
 end
 
-
-#--------------------------SUDOERS--------------------------#
-
-template "/etc/sudoers.d/redborder-manager" do
-  source "redborder-manager.erb"
-  owner "root"
-  group "root"
-  mode 0440
+# Sudoers
+template '/etc/sudoers.d/redborder-manager' do
+  source 'redborder-manager.erb'
+  owner 'root'
+  group 'root'
+  mode '0440'
   retries 2
 end
 
-#--------------------------Pending_changes--------------------------#
+# Pending Changes..
 # pending_changes==0 -> has changes to apply at next chef-client run
 #  pending_changes==1 -> chef-client has to run once
 #  pending_changes==2 -> chef-client has to run twice
 #  .......
 #  pending_changes==n -> chef-client has to run n times
 #
+node.normal['redborder']['pending_changes'] = node['redborder']['pending_changes'] > 0 ? node.normal['redborder']['pending_changes'].to_i - 1 : 0
 
-if node["redborder"]["pending_changes"]>0
-  node.normal["redborder"]["pending_changes"] = (node.normal["redborder"]["pending_changes"].to_i-1)
-else
-  node.normal["redborder"]["pending_changes"] = 0
-end
-
-execute "force_chef_client_wakeup" do
-  command "/usr/lib/redborder/bin/rb_wakeup_chef.sh"
+execute 'force_chef_client_wakeup' do
+  command '/usr/lib/redborder/bin/rb_wakeup_chef.sh'
   ignore_failure true
-  action ( node["redborder"]["pending_changes"].nil? or node["redborder"]["pending_changes"]==0 ) ? :nothing : :run
+  if node['redborder']['pending_changes'].nil? || node['redborder']['pending_changes'] == 0
+    action :nothing
+  else
+    action :run
+  end
 end
 
-#--------------------------MOTD--------------------------#
+# MOTD
+cluster_info = node['redborder']['cluster_info']
 
-cluster_info = node["redborder"]["cluster_info"]
-cluster_uuid_db = Chef::DataBagItem.load("rBglobal", "cluster") rescue cluster_uuid_db = {}
-cluster_installed = File.exist?("/etc/redborder/cluster-installed.txt")
+begin
+  cluster_uuid_db = data_bag_item('rBglobal', 'cluster')
+rescue
+  cluster_uuid_db = {}
+end
 
-template "/etc/motd" do
-    source "motd.erb"
-    owner "root"
-    group "root"
-    mode 0644
-    retries 2
-    backup false
-    variables(:cluster_info => cluster_info, :uuid => cluster_uuid_db["uuid"], :manager_services => manager_services, :cluster_finished => cluster_installed)
+cluster_installed = File.exist?('/etc/redborder/cluster-installed.txt')
+
+template '/etc/motd' do
+  source 'motd.erb'
+  owner 'root'
+  group 'root'
+  mode '0644'
+  retries 2
+  backup false
+  variables(cluster_info: cluster_info,
+            uuid: cluster_uuid_db['uuid'],
+            manager_services: manager_services,
+            cluster_finished: cluster_installed)
 end